requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab
XlsxWriter
email-validator
